Fix flaky tests (#23)
Summary:
Pull Request resolved: https://github.com/pytorch/elastic/pull/23

Test is flaky when run stress test. We did some investigation and found it due to port conflict, since we are using a const port in our tests, the port conflict can come from:
- same tests running in parallel in the same machine
- some other tests, or system using this static port

to fix this, we can find a free port before use it. (ideal way is run test in a container that everything is isolated, before that the approach mentioned above is a good workaround.)

Reviewed By: mehta-vikas

Differential Revision: D19173621

fbshipit-source-id: f2eef7ea83705cb85cc002462359428a4e874c2a