Extend agent functionality to work in heterogenous entironment
Summary:
The diff introduces several changes:
1. Change the way agents find out about the worker world size. Since in heterogenous env. the amount of workers can be different for each agent, we introduce a new algorithm for distributing the worker rank:

After the common store is set up between agents, we use it to advertise agents rank, role and amount of workers they have. Since we know the agent world size, each agent can retrieve the configurations of other agents. When the configuration is retrieved, the agent ranks can be assigned determenistically based on the position in the configuration array.

2. Introduce role based ranks. The agent's code is extended to work in the multi-role environment, where agents can have different role. We introduce two new environment variables: ROLE_RANK and ROLE_WORLD_SIZE. The ROLE_RANK variable is the proper worker rank distributed over the group of agents that have the same role.

Reviewed By: kiukchung

Differential Revision: D21611068

fbshipit-source-id: d51642e39e16fe8ddc161eb3fe8b241ab4b76fdd