Make torch elastic use new runtime_environment lib to abstract schedulers
Summary: Goal of this diff is to remove any scheduler-specific knowledge from TorchElastic and have it focus only on the parent-child process relationship for error reporting. This means that we should use the runtime_environment lib to interface with any scheduler-specific env vars, and unit tests should use mocks instead of setting scheduler-specific env vars.

Reviewed By: kiukchung

Differential Revision: D24661800

fbshipit-source-id: 14e3324d07fd4d6247ec1239ce1995b1fee41958