Make tsm api dataclasses serializable
Summary:
Make container to have a single resource object or a string representation of the resource instead of the Dict. The resources now resolved per scheduler during the drynrun method execution.

This is necessary to make the tsm APIs serializable, since resource capabilities can in general contain arbitrary non-serializable data.

Reviewed By: kiukchung

Differential Revision: D25877216

fbshipit-source-id: bfd8a9ccdaf826b1f353be126230167739e8ab19